finished updating the PHP models and frontend with the database changes for challenges the game can now be played as usual and the backend is ready to allow games of various game modes

+++ js/util.js
@@ -1,7 +1,7 @@
 function min(v1, v2)
 {
-    v1 = parseInt(v1);
-    v2 = parseInt(v2);
+    v1 = parseFloat(v1);
+    v2 = parseFloat(v2);
     if (isNaN(v1)) {
         return v2;
     }
@@ -13,8 +13,8 @@ function min(v1, v2)
 
 function max(v1, v2)
 {
-    v1 = parseInt(v1);
-    v2 = parseInt(v2);
+    v1 = parseFloat(v1);
+    v2 = parseFloat(v2);
     if (isNaN(v1)) {
         return v2;
     }
